Influence Measures and Diagnostic Plots for Multivariate Linear Models

This collection of functions is designed to compute regression deletion diagnostics for multivariate linear models following Barrett & Ling (1992). These are close analogs of standard methods for univariate and generalized linear models handled by the influence.measures in the stats package. These functions also extend plots of influence diagnostic measures such as those provided by influencePlot in the stats package.

In addition, the functions provide diagnostics for deletion of subsets of observations of size m>1. This case is theoretically interesting because sometimes pairs (m=2) of influential observations can mask each other, sometimes they can have joint influence far exceeding their individual effects, as well as other interesting phenomena described by Lawrence (1995). Associated methods for the case m>1 are still under development in this package.

Details

Package:	mvinfluence

Type:	Package

Version:	0.7

Date:	2013–9-06

License:	GPL-2

The design goal for this package is that, as an extension of standard methods for univariate linear models, you should be able to fit a linear model with a multivariate response,

  mymlm <- lm( cbind(y1, y2, y3) ~ x1 + x2 +x3, data=mydata)

and then get useful diagnostics and plots with

  influence(mymlm)
  hatvalues(mymlm)
  influencePlot(mymlm, ...)

Author

Michael Friendly

Maintainer: Michael Friendly <friendly@yorku.ca>

References

Barrett, B. E. and Ling, R. F. (1992). General Classes of Influence Measures for Multivariate Regression. Journal of the American Statistical Association, 87(417), 184-191.

Barrett, B. E. (2003). Understanding Influence in Multivariate Regression. Communications in Statistics – Theory and Methods, 32, 3, 667-680.

A. J. Lawrence (1995). Deletion Influence and Masking in Regression Journal of the Royal Statistical Society. Series B (Methodological) , Vol. 57, No. 1, pp. 181-189.

See also

Examples

# none here